% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Assignment}
\alias{Assignment}
\title{Assignment Object}
\usage{
Assignment(studentWorkFolder = NULL)
}
\arguments{
\item{studentWorkFolder}{Drive folder where attachments from student submissions are placed}
}
\value{
Assignment object
}
\description{
Assignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional details for assignments.
}
