% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{AssignmentSubmission}
\alias{AssignmentSubmission}
\title{AssignmentSubmission Object}
\usage{
AssignmentSubmission(attachments = NULL)
}
\arguments{
\item{attachments}{Attachments added by the student}
}
\value{
AssignmentSubmission object
}
\description{
AssignmentSubmission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Student work for an assignment.
}

