% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Attachment}
\alias{Attachment}
\title{Attachment Object}
\usage{
Attachment(driveFile = NULL, youTubeVideo = NULL, link = NULL,
  form = NULL)
}
\arguments{
\item{driveFile}{Google Drive file attachment}

\item{youTubeVideo}{Youtube video attachment}

\item{link}{Link attachment}

\item{form}{Google Forms attachment}
}
\value{
Attachment object
}
\description{
Attachment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Attachment added to student assignment work. When creating attachments, setting the `form` field is not supported.
}
