% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\docType{package}
\name{classroom_googleAuthR}
\alias{classroom_googleAuthR}
\alias{classroom_googleAuthR-package}
\title{Google Classroom API
Manages classes, rosters, and invitations in Google Classroom.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:56:19
filename: /Users/mark/dev/R/autoGoogleAPI/googleclassroomv1.auto/R/classroom_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/classroom.course-work.readonly
\item https://www.googleapis.com/auth/classroom.courses
\item https://www.googleapis.com/auth/classroom.courses.readonly
\item https://www.googleapis.com/auth/classroom.coursework.me
\item https://www.googleapis.com/auth/classroom.coursework.me.readonly
\item https://www.googleapis.com/auth/classroom.coursework.students
\item https://www.googleapis.com/auth/classroom.coursework.students.readonly
\item https://www.googleapis.com/auth/classroom.profile.emails
\item https://www.googleapis.com/auth/classroom.profile.photos
\item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.rosters.readonly
\item https://www.googleapis.com/auth/classroom.student-submissions.me.readonly
\item https://www.googleapis.com/auth/classroom.student-submissions.students.readonly
}
}
