% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Course}
\alias{Course}
\title{Google Classroom API Objects 
Manages classes, rosters, and invitations in Google Classroom.}
\usage{
Course(id = NULL, name = NULL, section = NULL,
  descriptionHeading = NULL, description = NULL, room = NULL,
  ownerId = NULL, creationTime = NULL, updateTime = NULL,
  enrollmentCode = NULL, courseState = NULL, alternateLink = NULL,
  teacherGroupEmail = NULL, courseGroupEmail = NULL, teacherFolder = NULL,
  courseMaterialSets = NULL, guardiansEnabled = NULL)
}
\arguments{
\item{id}{Identifier for this course assigned by Classroom}

\item{name}{Name of the course}

\item{section}{Section of the course}

\item{descriptionHeading}{Optional heading for the description}

\item{description}{Optional description}

\item{room}{Optional room location}

\item{ownerId}{The identifier of the owner of a course}

\item{creationTime}{Creation time of the course}

\item{updateTime}{Time of the most recent update to this course}

\item{enrollmentCode}{Enrollment code to use when joining this course}

\item{courseState}{State of the course}

\item{alternateLink}{Absolute link to this course in the Classroom web UI}

\item{teacherGroupEmail}{The email address of a Google group containing all teachers of the course}

\item{courseGroupEmail}{The email address of a Google group containing all members of the course}

\item{teacherFolder}{Information about a Drive Folder that is shared with all teachers of the course}

\item{courseMaterialSets}{Sets of materials that appear on the 'about' page of this course}

\item{guardiansEnabled}{Whether or not guardian notifications are enabled for this course}
}
\value{
Course object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:56:19
filename: /Users/mark/dev/R/autoGoogleAPI/googleclassroomv1.auto/R/classroom_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Course Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Course in Classroom.
}
\seealso{
Other Course functions: \code{\link{courses.create}},
  \code{\link{courses.patch}}, \code{\link{courses.update}}
}
