% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseMaterial}
\alias{CourseMaterial}
\title{CourseMaterial Object}
\usage{
CourseMaterial(driveFile = NULL, youTubeVideo = NULL, link = NULL,
  form = NULL)
}
\arguments{
\item{driveFile}{Google Drive file attachment}

\item{youTubeVideo}{Youtube video attachment}

\item{link}{Link atatchment}

\item{form}{Google Forms attachment}
}
\value{
CourseMaterial object
}
\description{
CourseMaterial Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A material attached to a course as part of a material set.
}
