% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseMaterialSet}
\alias{CourseMaterialSet}
\title{CourseMaterialSet Object}
\usage{
CourseMaterialSet(title = NULL, materials = NULL)
}
\arguments{
\item{title}{Title for this set}

\item{materials}{Materials attached to this set}
}
\value{
CourseMaterialSet object
}
\description{
CourseMaterialSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of materials that appears on the 'About' page of the course. These materials might include a syllabus, schedule, or other background information relating to the course as a whole.
}

