% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{CourseWork}
\alias{CourseWork}
\title{CourseWork Object}
\usage{
CourseWork(courseId = NULL, id = NULL, title = NULL, description = NULL,
  materials = NULL, state = NULL, alternateLink = NULL,
  creationTime = NULL, updateTime = NULL, dueDate = NULL,
  dueTime = NULL, maxPoints = NULL, workType = NULL,
  associatedWithDeveloper = NULL, submissionModificationMode = NULL,
  assignment = NULL, multipleChoiceQuestion = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{id}{Classroom-assigned identifier of this course work, unique per course}

\item{title}{Title of this course work}

\item{description}{Optional description of this course work}

\item{materials}{Additional materials}

\item{state}{Status of this course work}

\item{alternateLink}{Absolute link to this course work in the Classroom web UI}

\item{creationTime}{Timestamp when this course work was created}

\item{updateTime}{Timestamp of the most recent change to this course work}

\item{dueDate}{Optional date, in UTC, that submissions for this this course work are due}

\item{dueTime}{Optional time of day, in UTC, that submissions for this this course work are due}

\item{maxPoints}{Maximum grade for this course work}

\item{workType}{Type of this course work}

\item{associatedWithDeveloper}{Whether this course work item is associated with the Developer Console project making the request}

\item{submissionModificationMode}{Setting to determine when students are allowed to modify submissions}

\item{assignment}{Assignment details}

\item{multipleChoiceQuestion}{Multiple choice question details}
}
\value{
CourseWork object
}
\description{
CourseWork Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Course work created by a teacher for students of the course.
}

