% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{DriveFile}
\alias{DriveFile}
\title{DriveFile Object}
\usage{
DriveFile(id = NULL, title = NULL, alternateLink = NULL,
  thumbnailUrl = NULL)
}
\arguments{
\item{id}{Drive API resource ID}

\item{title}{Title of the Drive item}

\item{alternateLink}{URL that can be used to access the Drive item}

\item{thumbnailUrl}{URL of a thumbnail image of the Drive item}
}
\value{
DriveFile object
}
\description{
DriveFile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Representation of a Google Drive file.
}

