% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Form}
\alias{Form}
\title{Form Object}
\usage{
Form(formUrl = NULL, responseUrl = NULL, title = NULL,
  thumbnailUrl = NULL)
}
\arguments{
\item{formUrl}{URL of the form}

\item{responseUrl}{URL of the form responses document}

\item{title}{Title of the Form}

\item{thumbnailUrl}{URL of a thumbnail image of the Form}
}
\value{
Form object
}
\description{
Form Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Google Forms item.
}

