% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{GlobalPermission}
\alias{GlobalPermission}
\title{GlobalPermission Object}
\usage{
GlobalPermission(permission = NULL)
}
\arguments{
\item{permission}{Permission value}
}
\value{
GlobalPermission object
}
\description{
GlobalPermission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Global user permission description.
}

