% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_functions.R
\name{invitations.list}
\alias{invitations.list}
\title{Returns a list of invitations that the requesting user is permitted to view, restricted to those that match the list request. *Note:* At least one of `user_id` or `course_id` must be supplied. Both fields can be supplied. This method returns the following error codes: * `PERMISSION_DENIED` for access errors.}
\usage{
invitations.list(userId = NULL, courseId = NULL, pageSize = NULL,
  pageToken = NULL)
}
\arguments{
\item{userId}{Restricts returned invitations to those for a specific user}

\item{courseId}{Restricts returned invitations to those for a course with the specified identifier}

\item{pageSize}{Maximum number of items to return}

\item{pageToken}{nextPageToken value returned from a previous list call, indicating that the subsequent page of results should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/classroom.rosters
\item https://www.googleapis.com/auth/classroom.rosters.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/classroom.rosters, https://www.googleapis.com/auth/classroom.rosters.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/classroom/}{Google Documentation}
}
