% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Link}
\alias{Link}
\title{Link Object}
\usage{
Link(url = NULL, title = NULL, thumbnailUrl = NULL)
}
\arguments{
\item{url}{URL to link to}

\item{title}{Title of the target of the URL}

\item{thumbnailUrl}{URL of a thumbnail image of the target URL}
}
\value{
Link object
}
\description{
Link Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
URL item.
}
