% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListCourseAliasesResponse}
\alias{ListCourseAliasesResponse}
\title{ListCourseAliasesResponse Object}
\usage{
ListCourseAliasesResponse(aliases = NULL, nextPageToken = NULL)
}
\arguments{
\item{aliases}{The course aliases}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListCourseAliasesResponse object
}
\description{
ListCourseAliasesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing course aliases.
}
