% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListCoursesResponse}
\alias{ListCoursesResponse}
\title{ListCoursesResponse Object}
\usage{
ListCoursesResponse(courses = NULL, nextPageToken = NULL)
}
\arguments{
\item{courses}{Courses that match the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListCoursesResponse object
}
\description{
ListCoursesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing courses.
}
