% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListCourseWorkResponse}
\alias{ListCourseWorkResponse}
\title{ListCourseWorkResponse Object}
\usage{
ListCourseWorkResponse(courseWork = NULL, nextPageToken = NULL)
}
\arguments{
\item{courseWork}{Course work items that match the request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListCourseWorkResponse object
}
\description{
ListCourseWorkResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing course work.
}

