% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListGuardianInvitationsResponse}
\alias{ListGuardianInvitationsResponse}
\title{ListGuardianInvitationsResponse Object}
\usage{
ListGuardianInvitationsResponse(guardianInvitations = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{guardianInvitations}{Guardian invitations that matched the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListGuardianInvitationsResponse object
}
\description{
ListGuardianInvitationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing guardian invitations.
}

