% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListGuardiansResponse}
\alias{ListGuardiansResponse}
\title{ListGuardiansResponse Object}
\usage{
ListGuardiansResponse(guardians = NULL, nextPageToken = NULL)
}
\arguments{
\item{guardians}{Guardians on this page of results that met the criteria specified in the request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListGuardiansResponse object
}
\description{
ListGuardiansResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing guardians.
}

