% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListStudentsResponse}
\alias{ListStudentsResponse}
\title{ListStudentsResponse Object}
\usage{
ListStudentsResponse(students = NULL, nextPageToken = NULL)
}
\arguments{
\item{students}{Students who match the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListStudentsResponse object
}
\description{
ListStudentsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing students.
}

