% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListStudentSubmissionsResponse}
\alias{ListStudentSubmissionsResponse}
\title{ListStudentSubmissionsResponse Object}
\usage{
ListStudentSubmissionsResponse(studentSubmissions = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{studentSubmissions}{Student work that matches the request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListStudentSubmissionsResponse object
}
\description{
ListStudentSubmissionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing student submissions.
}
