% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ListTeachersResponse}
\alias{ListTeachersResponse}
\title{ListTeachersResponse Object}
\usage{
ListTeachersResponse(teachers = NULL, nextPageToken = NULL)
}
\arguments{
\item{teachers}{Teachers who match the list request}

\item{nextPageToken}{Token identifying the next page of results to return}
}
\value{
ListTeachersResponse object
}
\description{
ListTeachersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response when listing teachers.
}
