% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Material}
\alias{Material}
\title{Material Object}
\usage{
Material(driveFile = NULL, youtubeVideo = NULL, link = NULL,
  form = NULL)
}
\arguments{
\item{driveFile}{Google Drive file material}

\item{youtubeVideo}{YouTube video material}

\item{link}{Link material}

\item{form}{Google Forms material}
}
\value{
Material object
}
\description{
Material Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Material attached to course work. When creating attachments, only the Link field may be specified.
}

