% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ModifyAttachmentsRequest}
\alias{ModifyAttachmentsRequest}
\title{ModifyAttachmentsRequest Object}
\usage{
ModifyAttachmentsRequest(addAttachments = NULL)
}
\arguments{
\item{addAttachments}{Attachments to add}
}
\value{
ModifyAttachmentsRequest object
}
\description{
ModifyAttachmentsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to modify the attachments of a student submission.
}

