% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{MultipleChoiceQuestion}
\alias{MultipleChoiceQuestion}
\title{MultipleChoiceQuestion Object}
\usage{
MultipleChoiceQuestion(choices = NULL)
}
\arguments{
\item{choices}{Possible choices}
}
\value{
MultipleChoiceQuestion object
}
\description{
MultipleChoiceQuestion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional details for multiple-choice questions.
}

