% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{MultipleChoiceSubmission}
\alias{MultipleChoiceSubmission}
\title{MultipleChoiceSubmission Object}
\usage{
MultipleChoiceSubmission(answer = NULL)
}
\arguments{
\item{answer}{Student's select choice}
}
\value{
MultipleChoiceSubmission object
}
\description{
MultipleChoiceSubmission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Student work for a multiple-choice question.
}
