% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Name}
\alias{Name}
\title{Name Object}
\usage{
Name(givenName = NULL, familyName = NULL, fullName = NULL)
}
\arguments{
\item{givenName}{The user's first name}

\item{familyName}{The user's last name}

\item{fullName}{The user's full name formed by concatenating the first and last name values}
}
\value{
Name object
}
\description{
Name Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of the user's name.
}

