% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ReclaimStudentSubmissionRequest}
\alias{ReclaimStudentSubmissionRequest}
\title{ReclaimStudentSubmissionRequest Object}
\usage{
ReclaimStudentSubmissionRequest()
}
\value{
ReclaimStudentSubmissionRequest object
}
\description{
ReclaimStudentSubmissionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to reclaim a student submission.
}

