% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ReturnStudentSubmissionRequest}
\alias{ReturnStudentSubmissionRequest}
\title{ReturnStudentSubmissionRequest Object}
\usage{
ReturnStudentSubmissionRequest()
}
\value{
ReturnStudentSubmissionRequest object
}
\description{
ReturnStudentSubmissionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to return a student submission.
}
