% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{SharedDriveFile}
\alias{SharedDriveFile}
\title{SharedDriveFile Object}
\usage{
SharedDriveFile(driveFile = NULL, shareMode = NULL)
}
\arguments{
\item{driveFile}{Drive file details}

\item{shareMode}{Mechanism by which students access the Drive item}
}
\value{
SharedDriveFile object
}
\description{
SharedDriveFile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Drive file that is used as material for course work.
}
