% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{ShortAnswerSubmission}
\alias{ShortAnswerSubmission}
\title{ShortAnswerSubmission Object}
\usage{
ShortAnswerSubmission(answer = NULL)
}
\arguments{
\item{answer}{Student response to a short-answer question}
}
\value{
ShortAnswerSubmission object
}
\description{
ShortAnswerSubmission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Student work for a short answer question.
}
