% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Student}
\alias{Student}
\title{Student Object}
\usage{
Student(courseId = NULL, userId = NULL, profile = NULL,
  studentWorkFolder = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{userId}{Identifier of the user}

\item{profile}{Global user information for the student}

\item{studentWorkFolder}{Information about a Drive Folder for this student's work in this course}
}
\value{
Student object
}
\description{
Student Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Student in a course.
}

