% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{StudentSubmission}
\alias{StudentSubmission}
\title{StudentSubmission Object}
\usage{
StudentSubmission(courseId = NULL, courseWorkId = NULL, id = NULL,
  userId = NULL, creationTime = NULL, updateTime = NULL, state = NULL,
  late = NULL, draftGrade = NULL, assignedGrade = NULL,
  alternateLink = NULL, courseWorkType = NULL,
  associatedWithDeveloper = NULL, assignmentSubmission = NULL,
  shortAnswerSubmission = NULL, multipleChoiceSubmission = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{courseWorkId}{Identifier for the course work this corresponds to}

\item{id}{Classroom-assigned Identifier for the student submission}

\item{userId}{Identifier for the student that owns this submission}

\item{creationTime}{Creation time of this submission}

\item{updateTime}{Last update time of this submission}

\item{state}{State of this submission}

\item{late}{Whether this submission is late}

\item{draftGrade}{Optional pending grade}

\item{assignedGrade}{Optional grade}

\item{alternateLink}{Absolute link to the submission in the Classroom web UI}

\item{courseWorkType}{Type of course work this submission is for}

\item{associatedWithDeveloper}{Whether this student submission is associated with the Developer Console project making the request}

\item{assignmentSubmission}{Submission content when course_work_type is ASSIGNMENT}

\item{shortAnswerSubmission}{Submission content when course_work_type is SHORT_ANSWER_QUESTION}

\item{multipleChoiceSubmission}{Submission content when course_work_type is MULTIPLE_CHOICE_QUESTION}
}
\value{
StudentSubmission object
}
\description{
StudentSubmission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Student submission for course work. StudentSubmission items are generated when a CourseWork item is created. StudentSubmissions that have never been accessed (i.e. with `state` = NEW) may not have a creation time or update time.
}

