% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{Teacher}
\alias{Teacher}
\title{Teacher Object}
\usage{
Teacher(courseId = NULL, userId = NULL, profile = NULL)
}
\arguments{
\item{courseId}{Identifier of the course}

\item{userId}{Identifier of the user}

\item{profile}{Global user information for the teacher}
}
\value{
Teacher object
}
\description{
Teacher Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Teacher of a course.
}

