% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{TurnInStudentSubmissionRequest}
\alias{TurnInStudentSubmissionRequest}
\title{TurnInStudentSubmissionRequest Object}
\usage{
TurnInStudentSubmissionRequest()
}
\value{
TurnInStudentSubmissionRequest object
}
\description{
TurnInStudentSubmissionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to turn in a student submission.
}
