% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{UserProfile}
\alias{UserProfile}
\title{UserProfile Object}
\usage{
UserProfile(id = NULL, name = NULL, emailAddress = NULL,
  photoUrl = NULL, permissions = NULL)
}
\arguments{
\item{id}{Identifier of the user}

\item{name}{Name of the user}

\item{emailAddress}{Email address of the user}

\item{photoUrl}{URL of user's profile photo}

\item{permissions}{Global permissions of the user}
}
\value{
UserProfile object
}
\description{
UserProfile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Global information for a user.
}

