% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom_objects.R
\name{YouTubeVideo}
\alias{YouTubeVideo}
\title{YouTubeVideo Object}
\usage{
YouTubeVideo(id = NULL, title = NULL, alternateLink = NULL,
  thumbnailUrl = NULL)
}
\arguments{
\item{id}{YouTube API resource ID}

\item{title}{Title of the YouTube video}

\item{alternateLink}{URL that can be used to view the YouTube video}

\item{thumbnailUrl}{URL of a thumbnail image of the YouTube video}
}
\value{
YouTubeVideo object
}
\description{
YouTubeVideo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
YouTube video item.
}
